requirements.txt
freetype-py>=2.3.0
numpy>=1.20.0
tqdm>=4.60.0
Pillow>=8.0.0
